refactor: 첨부파일 로직을 useAttachmentHandler 훅으로 분리
