+++ src/screens/chat/ChatRoomPage.tsx
@@ -15,8 +15,8 @@ import {
 import Image from 'next/image';
 import { useRouter, useSearchParams } from 'next/navigation';
 import {
+  startTransition,
   useCallback,
-  type ChangeEvent,
   useEffect,
   useLayoutEffect,
   useMemo,
@@ -29,25 +29,20 @@ import ConfirmModal from '@/components/common/ConfirmModal';
 import { useAppFrame } from '@/components/layout/AppFrameContext';
 import { useHeader } from '@/components/layout/HeaderContext';
 import {
-  ALLOWED_FILE_MIME_TYPES,
-  ALLOWED_IMAGE_MIME_TYPES,
   BOTTOM_CONFIRM_THRESHOLD,
   DELETE_LONG_PRESS_MS,
   LONG_MESSAGE_THRESHOLD,
-  MAX_ATTACHMENT_FILE_SIZE_BYTES,
-  MAX_IMAGE_ATTACHMENTS_PER_PICK,
   MESSAGE_PAGE_SIZE,
   MESSAGE_SEND_DESTINATION,
-  PDF_FILE_EXTENSION_REGEX,
   TOP_FETCH_THRESHOLD,
 } from '@/constants/chat';
-import { fetchChatMessages } from '@/lib/api/chatMessages';
 import { getUserIdFromAccessToken } from '@/lib/auth/token';
 import { applyRealtimeRoomMessage } from '@/lib/chat/realtimeMessageCache';
 import { applyRealtimeRoomNotification } from '@/lib/chat/realtimeRoomCache';
 import { clearRejoinedRoomUiOverride } from '@/lib/chat/rejoinedRoomUiCache';
 import { chatStompManager } from '@/lib/chat/stompManager';
 import { chatKeys } from '@/lib/hooks/chat/queryKeys';
+import { useAttachmentHandler } from '@/lib/hooks/chat/useAttachmentHandler';
 import { useChatMessagesInfiniteQuery } from '@/lib/hooks/chat/useChatMessagesInfiniteQuery';
 import { useChatRoomDetailQuery } from '@/lib/hooks/chat/useChatRoomDetailQuery';
 import { useChatSubscriptions } from '@/lib/hooks/chat/useChatSubscriptions';
@@ -56,7 +51,6 @@ import { useLeaveChatRoomMutation } from '@/lib/hooks/chat/useLeaveChatRoomMutat
 import { usePatchLastReadMutation } from '@/lib/hooks/chat/usePatchLastReadMutation';
 import { usePutRoomSettingsMutation } from '@/lib/hooks/chat/usePutRoomSettingsMutation';
 import { toast } from '@/lib/toast/store';
-import { uploadFile } from '@/lib/upload/uploadFile';
 import { formatDateKey, formatMessageTime, formatStickyDateLabel } from '@/lib/utils/chatDate';
 
 import type { ChatMessageResponse, SendChatMessagePayload } from '@/lib/api/chatMessages';
@@ -66,24 +60,6 @@ type ChatRoomPageProps = Readonly<{
   roomId: number | null;
   mode?: 'room' | 'settings';
 }>;
-
-function isPdfFile(file: File): boolean {
-  const mimeType = file.type.trim().toLowerCase();
-  if (mimeType && ALLOWED_FILE_MIME_TYPES.has(mimeType)) {
-    return true;
-  }
-
-  return PDF_FILE_EXTENSION_REGEX.test(file.name);
-}
-
-function resolveAttachmentMimeType(file: File, isImageAttachment: boolean): string {
-  if (isImageAttachment) {
-    return file.type || 'application/octet-stream';
-  }
-
-  // Some environments return empty/variant MIME for PDF. Normalize to application/pdf.
-  return isPdfFile(file) ? 'application/pdf' : file.type || 'application/octet-stream';
-}
 
 function resolveChatAssetUrl(s3KeyOrUrl: string | null): string | null {
   if (!s3KeyOrUrl) {
@@ -157,11 +133,6 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   const [isLeavingRoom, setIsLeavingRoom] = useState(false);
   const [roomNameInput, setRoomNameInput] = useState('');
   const [isAlarmOnInput, setIsAlarmOnInput] = useState(true);
-  const [isAttachmentUploading, setIsAttachmentUploading] = useState(false);
-  const [isAttachmentPickerOpen, setIsAttachmentPickerOpen] = useState(false);
-  const [attachmentValidationMessage, setAttachmentValidationMessage] = useState<string | null>(
-    null,
-  );
   const [imagePreview, setImagePreview] = useState<{
     src: string;
     alt: string;
@@ -170,8 +141,6 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   const activeRoomId = isLeavingRoom ? null : roomId;
   const { data, isLoading, isError, refetch } = useChatRoomDetailQuery(activeRoomId);
   const messageListRef = useRef<HTMLDivElement>(null);
-  const imageAttachmentInputRef = useRef<HTMLInputElement>(null);
-  const fileAttachmentInputRef = useRef<HTMLInputElement>(null);
   const unreadDividerRef = useRef<HTMLDivElement>(null);
   const deleteLongPressTimerRef = useRef<number | null>(null);
   const isMessageInputComposingRef = useRef(false);
@@ -182,7 +151,6 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   const prevScrollHeightRef = useRef(0);
   const hasPatchedOnEntryRef = useRef(false);
   const lastPatchedMsgIdRef = useRef<number | null>(null);
-  const attachmentEchoFallbackTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
   const patchLastReadMutation = usePatchLastReadMutation(roomId ?? 0);
   const deleteMessageMutation = useDeleteMessageMutation(roomId ?? 0);
   const putRoomSettingsMutation = usePutRoomSettingsMutation(roomId ?? 0);
@@ -234,6 +202,20 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   }, [currentUserId, messages, serverLastReadMsgId]);
 
   const isPrivateRoom = data?.type === 'PRIVATE';
+
+  const {
+    isAttachmentUploading,
+    isAttachmentPickerOpen,
+    setIsAttachmentPickerOpen,
+    attachmentValidationMessage,
+    setAttachmentValidationMessage,
+    imageAttachmentInputRef,
+    fileAttachmentInputRef,
+    handleAttachmentButtonClick,
+    handlePickImageAttachments,
+    handlePickFileAttachment,
+    handleAttachmentChange,
+  } = useAttachmentHandler({ roomId, queryClient });
 
   const toggleExpandedMessage = useCallback((messageId: number) => {
     setExpandedMessageIds((prev) => {
@@ -350,147 +332,6 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
       setMessageInput('');
     },
     [messageInput, roomId],
-  );
-
-  const handleAttachmentButtonClick = useCallback(() => {
-    if (isAttachmentUploading) {
-      return;
-    }
-
-    setIsAttachmentPickerOpen(true);
-  }, [isAttachmentUploading]);
-
-  const openAttachmentValidationModal = useCallback((message: string) => {
-    setIsAttachmentPickerOpen(false);
-    setAttachmentValidationMessage(message);
-  }, []);
-
-  const handlePickImageAttachments = useCallback(() => {
-    if (isAttachmentUploading) {
-      return;
-    }
-
-    setIsAttachmentPickerOpen(false);
-    imageAttachmentInputRef.current?.click();
-  }, [isAttachmentUploading]);
-
-  const handlePickFileAttachment = useCallback(() => {
-    if (isAttachmentUploading) {
-      return;
-    }
-
-    setIsAttachmentPickerOpen(false);
-    fileAttachmentInputRef.current?.click();
-  }, [isAttachmentUploading]);
-
-  const handleAttachmentChange = useCallback(
-    async (event: ChangeEvent<HTMLInputElement>) => {
-      const selectedFiles = Array.from(event.target.files ?? []);
-      event.target.value = '';
-
-      if (selectedFiles.length === 0 || roomId === null || isAttachmentUploading) {
-        return;
-      }
-
-      const imageFiles = selectedFiles.filter((file) => ALLOWED_IMAGE_MIME_TYPES.has(file.type));
-      const nonImageFiles = selectedFiles.filter(
-        (file) => !ALLOWED_IMAGE_MIME_TYPES.has(file.type),
-      );
-
-      if (imageFiles.length > 0 && nonImageFiles.length > 0) {
-        openAttachmentValidationModal('이미지와 파일은 동시에 첨부할 수 없습니다.');
-        return;
-      }
-
-      if (imageFiles.length > MAX_IMAGE_ATTACHMENTS_PER_PICK) {
-        openAttachmentValidationModal(
-          `이미지는 한 번에 최대 ${MAX_IMAGE_ATTACHMENTS_PER_PICK}장까지 첨부할 수 있습니다.`,
-        );
-        return;
-      }
-
-      if (nonImageFiles.length > 1) {
-        openAttachmentValidationModal('파일은 한 번에 1개만 첨부할 수 있습니다.');
-        return;
-      }
-
-      for (const file of selectedFiles) {
-        if (file.size > MAX_ATTACHMENT_FILE_SIZE_BYTES) {
-          openAttachmentValidationModal('파일 용량은 5MB 이하만 첨부할 수 있습니다.');
-          return;
-        }
-      }
-
-      if (nonImageFiles.length === 1 && !isPdfFile(nonImageFiles[0]!)) {
-        openAttachmentValidationModal('파일 첨부는 PDF 형식만 지원합니다.');
-        return;
-      }
-
-      setIsAttachmentUploading(true);
-
-      try {
-        for (const file of selectedFiles) {
-          const isImageAttachment = ALLOWED_IMAGE_MIME_TYPES.has(file.type);
-          const normalizedMimeType = resolveAttachmentMimeType(file, isImageAttachment);
-
-          const uploaded = await uploadFile({
-            file,
-            category: 'AI_CHAT_ATTACHMENT',
-            refType: 'CHATROOM',
-            refId: roomId,
-            mimeType: normalizedMimeType,
-          });
-
-          const payload: SendChatMessagePayload = isImageAttachment
-            ? {
-                roomId,
-                type: 'IMAGE',
-                content: null,
-                s3Key: uploaded.s3Key,
-              }
-            : {
-                roomId,
-                type: 'PDF',
-                content: uploaded.s3Key,
-                s3Key: uploaded.s3Key,
-              };
-
-          const published = chatStompManager.publishJson(MESSAGE_SEND_DESTINATION, payload);
-          if (!published) {
-            toast('첨부 메시지 전송에 실패했습니다. 잠시 후 다시 시도해 주세요.');
-            return;
-          }
-
-          // STOMP echo fallback: 백엔드가 FILE/IMAGE 메시지를 브로드캐스트하지 않는 경우
-          // 2초 후 REST API로 최신 메시지를 가져와 캐시에 병합합니다.
-          if (attachmentEchoFallbackTimerRef.current !== null) {
-            clearTimeout(attachmentEchoFallbackTimerRef.current);
-          }
-          const fallbackRoomId = roomId;
-          attachmentEchoFallbackTimerRef.current = setTimeout(() => {
-            attachmentEchoFallbackTimerRef.current = null;
-            void fetchChatMessages(fallbackRoomId, { size: MESSAGE_PAGE_SIZE }).then((result) => {
-              if (!result.ok || !result.json || !('data' in result.json) || !result.json.data) {
-                return;
-              }
-              for (const msg of result.json.data.messages) {
-                applyRealtimeRoomMessage(queryClient, {
-                  roomId: fallbackRoomId,
-                  size: MESSAGE_PAGE_SIZE,
-                  message: msg,
-                });
-              }
-            });
-          }, 2000);
-        }
-      } catch (error) {
-        const err = error as Error;
-        toast(err.message || '파일 업로드에 실패했습니다.');
-      } finally {
-        setIsAttachmentUploading(false);
-      }
-    },
-    [isAttachmentUploading, openAttachmentValidationModal, queryClient, roomId],
   );
 
   useChatSubscriptions({
@@ -663,9 +504,9 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   ]);
 
   useEffect(() => {
-    if (attachmentEchoFallbackTimerRef.current !== null) {
-      clearTimeout(attachmentEchoFallbackTimerRef.current);
-      attachmentEchoFallbackTimerRef.current = null;
+    if (initialScrollResyncTimerRef.current !== null) {
+      window.clearTimeout(initialScrollResyncTimerRef.current);
+      initialScrollResyncTimerRef.current = null;
     }
     hasInitialScrollRef.current = false;
     hasNoUnreadInitialBottomResyncedRef.current = false;
@@ -673,8 +514,10 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
     prevScrollHeightRef.current = 0;
     hasPatchedOnEntryRef.current = false;
     lastPatchedMsgIdRef.current = null;
-    setIsLeavingRoom(false);
-    setMessageInput('');
+    startTransition(() => {
+      setIsLeavingRoom(false);
+      setMessageInput('');
+    });
   }, [roomId]);
 
   useEffect(() => {
@@ -689,8 +532,10 @@ export default function ChatRoomPage({ roomId, mode = 'room' }: ChatRoomPageProp
   }, [queryClient, roomId]);
 
   useEffect(() => {
-    setRoomNameInput(data?.roomName ?? '');
-    setIsAlarmOnInput(data?.isAlarmOn ?? true);
+    startTransition(() => {
+      setRoomNameInput(data?.roomName ?? '');
+      setIsAlarmOnInput(data?.isAlarmOn ?? true);
+    });
   }, [data?.isAlarmOn, data?.roomName, roomId]);
 
   useEffect(() => {
